requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.99.2
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy==2.0.36
passlib==1.7.4
requests==2.32.3
anytree==2.12.1
PyJWT==2.8.0
python-multipart>=0.0.9